### 2025-05-15 #8204 @editorialbot
[PRE REVIEW]: smolyax: a high-performance implementation of the Smolyak interpolation operator in JAX
**Submitting author:** @JoWestermann (<a href="http://orcid.org/0000-0003-3450-9166">Josephine</a>)
**Repository:** https://github.com/JoWestermann/smolyax
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
