### 2025-05-15 #8204 @danielskatz
👋 @JoWestermann - thanks for your submission.

I see two issues:
1. Due to the relatively small amount of code, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. You should hear back in a week or two.
2. If this does go ahead, you will need to reduce the length of the paper. JOSS's guidelines are 250-1000 words, while your paper is about 1750.  In case this doesn't pass our scope review, I don't think you should start reducing the paper yet.
