### 2025-05-17 #8204 @JoWestermann
Dear @danielskatz, thank you for your this first feedback. In case this might tip the scale in terms of lines of code, we do have several additional features sketched out or in the works that would expand the code base further, and which we could add over the course of a few weeks.
We’re looking forward to hearing back from you and the editors.
