### 2025-05-26 15c50f3
Moved compute_b functionality to barycentric module, added docstrings and typehints
