### 2025-05-29 c9d0625
Moved evaluate_tensor_product_interpolant to barycentric
