### 2025-05-29 e4bb391
Minimal refactoring of init_nodes_and_weights
