### 2025-05-29 9850ad7
Rename __compiledfuncs to __compiled_tensor_product_evaluation
