### 2025-05-29 a0524b9
Quadrature draft (not working yet)
