### 2025-05-31 845425c
Added ipykernel and matplotlib to [dev] dependencies
