### 2025-05-31 e515d6e
Refactored setup of test polynomial
