+++ tests/test_smolyak.py
@@ -15,7 +15,7 @@ def test_interpolation():
         t = np.random.uniform(low=1, high=8)
         print(f"... with k = {k}, t = {t}, d_out = {d_out},", node_gen)
 
-        f = setup.generate_test_function_smolyak(node_gen=node_gen, k=k, t=t, d_out=d_out)
+        f = setup.TestPolynomial(node_gen=node_gen, k=k, t=t, d_out=d_out)
 
         ip = SmolyakBarycentricInterpolator(node_gen=node_gen, k=k, t=t, d_out=d_out, batchsize=1, f=f)
 
@@ -37,7 +37,7 @@ def test_quadrature():
         t = np.random.uniform(low=1, high=8)
         print(f"... with k = {k}, t = {t}, d_out = {d_out},", node_gen)
 
-        f = setup.generate_test_function_smolyak(node_gen=node_gen, k=k, t=t, d_out=d_out)
+        f = setup.TestPolynomial(node_gen=node_gen, k=k, t=t, d_out=d_out)
 
         ip = SmolyakBarycentricInterpolator(node_gen=node_gen, k=k, t=t, d_out=d_out, batchsize=1, f=f)
 
