### 2025-06-04 45c188e
Get quadrature working on default domains
