### 2025-06-06 d7656fb
Get quadrature working on arbitrary domains
