+++ tests/test_smolyak.py
@@ -29,7 +29,7 @@ def test_interpolation():
 def test_quadrature():
     print("\nTesting vector-valued Smolyak quadrature ...")
 
-    for node_gen in setup.generate_nodes_default(dmin=1, dmax=6):
+    for node_gen in setup.generate_nodes(n=20, dmin=1, dmax=4):
 
         k = sorted(np.random.uniform(low=1, high=10, size=node_gen.dim))
         k /= k[0]
