### 2025-06-06 f51943a
Added gradient functionality
