### 2025-06-11 f3eea4f
Use caching for quadrature weights
