### 2025-06-11 a49c320
Added new features to readme
