### 2025-06-11 25ffa48
Added documentation for gradient and integral
