### 2025-06-13 1bec759
Quadrature (#45)
* Moved compute_b functionality to barycentric module, added docstrings and typehints

* Updated documentation (including leftover change jax_smolyak -> smolyax)

* Moved evaluate_tensor_product_interpolant to barycentric

* Minimal refactoring of init_nodes_and_weights

* Rename __compiledfuncs to __compiled_tensor_product_evaluation

* Quadrature draft (not working yet)

* Added ipykernel and matplotlib to [dev] dependencies

* Quadrature fix

* Refactored setup of test polynomial

* Get quadrature working on default domains

* Get quadrature working on arbitrary domains

* 
[…]
