### 2025-06-13 4857e25
Set default batchsize to None, which skips warm up call
