### 2025-06-15 65000a3
Remove distinction between centered and non-centered domains in evaluation of barycentric basis since it no longer offers consistent performance benefit
