+++ src/smolyax/barycentric.py
@@ -2,7 +2,7 @@
 Utility functionality for barycentric interpolation.
 """
 
-from typing import Callable, Sequence, Union
+from typing import Sequence, Union
 
 import jax
 import jax.numpy as jnp
@@ -30,42 +30,9 @@ def compute_weights(nodes: Union[jax.Array, np.ndarray]) -> jax.Array:
     return jnp.prod(1 / diffs, axis=0)
 
 
-def evaluate_basis_numerator_centered(x: jax.Array, xi: jax.Array, w: jax.Array, nu_i: int) -> jax.Array:
+def evaluate_basis_unnormalized(x: jax.Array, xi: jax.Array, w: jax.Array, nu_i: int) -> jax.Array:
     r"""
-    Evaluate the barycentric basis numerator terms at given evaluation points (assuming a centered domain).
-
-    Computes the vector of terms $w_j / (x - \xi_j)$ for a single univariate interpolant.
-    If an evaluation point coincides with a node, returns a one-hot indicator pattern.
-    All entries of xi and w beyond index `nu_i` are assumed to be zero (which is not checked for efficiency).
-
-    Parameters
-    ----------
-    x : jax.Array
-        Evaluation points along one dimension. Shape: `(n_points, 1)`
-    xi : jax.Array
-        Interpolation nodes. Shape `(m_i,)` with `m_i > nu_i`. Entries beyond `nu_i` are assumed to be zero.
-    w : jax.Array
-        Barycentric weights corresponding to the given interpolation nodes. Shape `(m_i,)` with `m_i > nu_i`.
-        Entries beyond `nu_i` are assumed to be zero.
-    nu_i : int
-        Polynomial degree in this dimension (unused here, included for interface compatibility).
-
-    Returns
-    -------
-    b : jax.Array
-        Barycentric numerator terms $w_j / (x - x_j)$, or a one-hot indicator pattern if `x` coincides with a node.
-        Shape `(n_points, m_i)`.
-    """
-    diffs = x - xi
-    mask_zero = jnp.any(diffs == 0, axis=1)
-    one_hot_pattern = jnp.where(diffs == 0, 1.0, 0.0)
-    w_div_diffs = w / diffs
-    return jnp.where(mask_zero[:, None], one_hot_pattern, w_div_diffs)
-
-
-def evaluate_basis_numerator_noncentered(x: jax.Array, xi: jax.Array, w: jax.Array, nu_i: int) -> jax.Array:
-    r"""
-    Evaluate the barycentric basis numerator terms at given evaluation points (assuming a non-centered domain).
+    Evaluate the barycentric basis numerator terms at given evaluation points.
 
     Computes $w_j / (x - \xi_j)$ for indices $j \le \nu_i$; other entries are masked to zero.
     If an evaluation point coincides with a node $\xi_j$ for $j \le \nu_i$, returns a one-hot indicator pattern
@@ -100,7 +67,6 @@ def evaluate_basis_numerator_noncentered(x: jax.Array, xi: jax.Array, w: jax.Arr
 
 def evaluate_tensor_product_interpolant(
     x: jax.Array,
-    evaluate_basis_numerator: Callable,
     F: jax.Array,
     xi_list: Sequence[jax.Array],
     w_list: Sequence[jax.Array],
@@ -116,10 +82,6 @@ def evaluate_tensor_product_interpolant(
         Points at which to evaluate the tensor product interpolant of the target function `f`.
         Should be a 2D array of shape `(n_points, d_in)` where `n_points` is the number of evaluation points
         and `d_in` is the dimension of the input domain.
-
-    evaluate_basis_numerator : Callable
[…]
