### 2025-06-15 f32f61e
Implement jacobian via analytic formula instead of autodiff
