+++ src/smolyax/barycentric.py
@@ -74,7 +74,7 @@ def evaluate_tensor_product_interpolant(
     sorted_degs: Sequence[int],
 ) -> jax.Array:
     """
-    Evaluate a tensor product interpolant using the barycentric formulation.
+    Evaluate the tensor product interpolant using the barycentric formulation.
 
     Parameters
     ----------
@@ -115,3 +115,107 @@ def evaluate_tensor_product_interpolant(
             F = jnp.einsum("ij,ikj...->ik...", b, F)
         norm *= jnp.sum(b, axis=1)
     return F / norm[:, None]
+
+
+def evaluate_basis_gradient_unnormalized(x: jax.Array, xi: jax.Array, w: jax.Array, nu_i: int) -> jax.Array:
+    r"""
+    Evaluate the gradient of the barycentric basis numerator terms at given evaluation points.
+
+    Computes $w_j / (x - \xi_j)$ for indices $j \le \nu_i$; other entries are masked to zero.
+    If an evaluation point coincides with a node $\xi_j$ for $j \le \nu_i$, returns a one-hot indicator pattern
+    restricted to those entries.
+
+    Parameters
+    ----------
+    x : jax.Array
+        Evaluation points along one dimension. Shape: `(n_points, 1)`
+    xi : jax.Array
+        Interpolation nodes. Shape `(m_i,)` with `m_i > nu_i`. Entries beyond `nu_i` are not used.
+    w : jax.Array
+        Barycentric weights corresponding to the given interpolation nodes. Shape `(m_i,)` with `m_i > nu_i`.
+        Entries beyond `nu_i` are not used.
+    nu_i : int
+        Polynomial degree in this dimension.
+
+    Returns
+    -------
+    b : jax.Array
+        Gradient of the barycentric numerator terms for indices $j \le \nu_i$, or a one-hot indicator pattern if `x`
+        coincides with a node. Shape `(n_points, m_i)`.
+    """
+    diffs = (x - xi) ** 2
+    mask_cols = jnp.arange(diffs.shape[1]) <= nu_i
+    mask_zero = jnp.any((diffs == 0) & mask_cols, axis=1)
+    one_hot_pattern = (diffs == 0).astype(diffs.dtype)
+    w_div_diffs = jnp.divide(w, diffs)
+    b = jnp.where(mask_zero[:, None], one_hot_pattern, w_div_diffs)
+    return jnp.where(mask_cols[None, :], b, 0)
+
+
+def evaluate_tensor_product_gradient(
+    x: jax.Array,
+    F: jax.Array,
+    xi_list: Sequence[jax.Array],
+    w_list: Sequence[jax.Array],
+    sorted_dims: Sequence[int],
+    sorted_degs: Sequence[int],
+) -> jax.Array:
+    """
+    Evaluate the gradient of the tensor product interpolant using the barycentric formulation.
+
+    Parameters
+    ----------
+    x : jax.Array
+        Points at which to evaluate the tensor product interpolant of the target function `f`.
+        Should be a 2D array of shape `(n_points, d_in)` where `n_points` is the number of evaluation points
+        and `d_in` is the dimension of the input domain.
+
+    F : jax.Array
+        Tensors storing the evaluations of the target function `f`.
+        Should be a multi-dimensional array with shape `(d_out, mu_1, mu_2, ..., mu_n)`
+        where each `mu_i` corresponds to the number of points in the ith dimension.
+
+    xi_list : Sequence[jax.Array]
+        Interpolation nodes. A sequence of 1D arrays, each with shape `(mu_i,)` for the ith dimension.
+
+    w_list : Sequence[jax.Array]
[…]
