### 2025-06-16 63af312
Clearer separation between 1D node sequence generators and container class, along with improved documentation
