+++ src/smolyax/nodes/base.py
@@ -1,88 +1,237 @@
 from abc import ABC, abstractmethod
-from typing import Iterator, List, Self, Union
+from typing import Iterator, List, Union
 
 import jax
 import numpy as np
 
 
-class Generator(ABC):
-    """Abstract base class for one-dimensional interpolation nodes"""
+class Generator1D(ABC):
+    """
+    Abstract base class for one-dimensional node sequences.
 
-    def __init__(self, dim: int, is_nested: bool) -> None:
+    These sequences can be used for interpolation or quadrature. Nodes are typically
+    defined on a reference domain, but `scale` and `scale_back` allow mapping to custom domains.
+    Supports random sampling from the measure of integration associated with the domain.
+    """
+
+    def __init__(self, is_nested: bool) -> None:
         """
-        Initialize the generator.
+        Initialize the node sequence generator.
 
         Parameters
         ----------
-        dim : int
-            Dimensionality of the generator.
         is_nested : bool
-            Indicates if the node sequence is nested.
+            Whether the node sequence is nested.
         """
-        self.__dim = dim
         self.__is_nested = is_nested
 
     @property
     def is_nested(self) -> bool:
         """
-        `True` if the node sequence is nested, `False` otherwise.
+        Whether the node sequence is nested.
+
+        Returns
+        -------
+        bool
         """
         return self.__is_nested
+
+    @abstractmethod
+    def __call__(self, n: int) -> Union[jax.Array, np.ndarray]:
+        """@public
+        Generate a node sequence of length `n+1`.
+
+        Parameters
+        ----------
+        n : int
+            Maximal node index (counting starts at `0`).
+
+        Returns
+        -------
+        jax.Array or np.ndarray
+            The generated node sequence.
+        """
+        ...
+
+    @abstractmethod
+    def scale(self, x: Union[jax.Array, np.ndarray]) -> Union[jax.Array, np.ndarray]:
+        """
+        Map nodes and points from the reference domain to a custom domain.
+
+        Mainly intended for internal use and testing.
+
+        Parameters
+        ----------
+        x : jax.Array or np.ndarray
+            Nodes or points on the reference domain.
+
+        Returns
+        -------
+        jax.Array or np.ndarray
+            Nodes or points mapped to the custom domain.
+        """
+        ...
+
+    @abstractmethod
+    def scale_back(self, x: Union[jax.Array, np.ndarray]) -> Union[jax.Array, np.ndarray]:
+        """
+        Map nodes and points from the custom domain back to the reference domain.
+
+        Mainly intended for internal use and testing.
+
+        Parameters
+        ----------
+        x : jax.Array or np.ndarray
+            Nodes or points on the custom domain.
+
+        Returns
+        -------
+        jax.Array or np.ndarray
+            Nodes or points mapped to the reference domain.
+        """
+        ...
+
+    @abstractmethod
+    def get_random(self, n: int = 1) -> Union[jax.Array, np.ndarray]:
+        """
+        Sample points randomly from the measure associated with the domain.
+
+        This measure corresponds, for example, to the integration measure for quadrature.
+
+        Parameters
+        ----------
+        n : int, optional
+            Number of random points to sample (default is 1).
+
+        Returns
+        -------
+        jax.Array or np.ndarray
+            Randomly sampled points.
+        """
+        ...
+
+    @abstractmethod
+    def get_quadrature_weights(self, n: int) -> Union[jax.Array, np.ndarray]:
+        """
+        Return the quadrature weights for the node sequence.
+
+        Parameters
[…]
