+++ src/smolyax/nodes/leja.py
@@ -4,15 +4,15 @@ from typing import Sequence, Union
 import jax
 import numpy as np
 
-from .base import Generator, GeneratorMultiD
+from .base import Generator, Generator1D
 
 
-class Leja1D(Generator):
-    """A generator for Leja points, a nested sequence on a closed one-dimensional domain."""
+class Leja1D(Generator1D):
+    """
+    Generator for one-dimensional Leja nodes.
 
-    @property
-    def domain(self) -> Union[jax.Array, np.ndarray, Sequence[float]]:
-        return self.__domain
+    Produces nested Leja nodes (default on [-1, 1]), with optional scaling to arbitrary bounded domains.
+    """
 
     __nodes = np.array([0, 1, -1, 1 / np.sqrt(2), -1 / np.sqrt(2)])
 
@@ -25,7 +25,7 @@ class Leja1D(Generator):
         domain : Union[jax.Array, np.ndarray, Sequence[float]], optional
             Endpoints of the domain. Defaults to `[-1, 1]` if not specified.
         """
-        super().__init__(dim=1, is_nested=True)
+        super().__init__(is_nested=True)
         self.__domain = domain
         self.__reference_domain = None
         if domain is not None:
@@ -34,8 +34,16 @@ class Leja1D(Generator):
         self.__cached_call = self.__make_cached_call()
         self.__cached_get_quadrature_weights = self.__make_cached_get_quadrature_weights()
 
-    def __repr__(self) -> str:
-        return f"Leja (domain = {self.__domain})"
+    @property
+    def domain(self) -> Union[jax.Array, np.ndarray, Sequence[float]]:
+        """
+        Domain interval of the node sequence.
+
+        Returns
+        -------
+        array-like of shape (2,)
+        """
+        return self.__domain
 
     @classmethod
     def __ensure_nodes(cls, n: int):
@@ -57,6 +65,19 @@ class Leja1D(Generator):
         return cached
 
     def __call__(self, n: int) -> Union[jax.Array, np.ndarray]:
+        """
+        Generate Leja nodes mapped to `self.domain`.
+
+        Parameters
+        ----------
+        n : int
+            Maximal node index (counting starts at `0`).
+
+        Returns
+        -------
+        jax.Array or np.ndarray
+            Leja nodes mapped to the `self.domain`.
+        """
         return self.__cached_call(n)
 
     def __make_cached_get_quadrature_weights(self):
@@ -71,9 +92,6 @@ class Leja1D(Generator):
 
         return cached
 
-    def get_quadrature_weights(self, n: int) -> Union[jax.Array, np.ndarray]:
-        return self.__cached_get_quadrature_weights(n)
-
     def scale(
         self,
         x: Union[jax.Array, np.ndarray],
@@ -81,9 +99,23 @@ class Leja1D(Generator):
         d2: Union[jax.Array, np.ndarray, Sequence[float]] = None,
     ) -> Union[jax.Array, np.ndarray]:
         """
-        Affine transformation from the interval d1 to the interval d2 applied to point x.
-        x : scalar or array or list of shape (n, )
-        d1, d2 : arrays or lists of shape (2, )
+        Map nodes or points from interval `d1` to interval `d2` by affine transformation.
+
+        Mainly intended for internal use and testing.
+
+        Parameters
+        ----------
+        x : array-like
+            Points to transform.
+        d1 : array-like of shape (2,), optional
+            Source interval (default: reference domain `[-1, 1]`).
+        d2 : array-like of shape (2,), optional
+            Target interval (default: generator's domain).
+
+        Returns
+        -------
+        jax.Array or np.ndarray
+            Nodes or points in the target domain.
         """
         if d1 is None:
             d1 = self.__reference_domain
@@ -119,14 +151,63 @@ class Leja1D(Generator):
[…]
