+++ tests/setup.py
@@ -48,7 +48,7 @@ def generate_nodes(*, n: int, dmin: int, dmax: int) -> List[nodes.Generator]:
     return sets
 
 
-def evaluate_univariate_polynomial(node_gen_uni: nodes.Generator, degree: int, x: np.ndarray) -> ArrayLike:
+def evaluate_univariate_polynomial(node_gen_uni: nodes.Generator1D, degree: int, x: np.ndarray) -> ArrayLike:
     x = node_gen_uni.scale_back(x)
     coefficients = [0] * degree + [1]
     if isinstance(node_gen_uni, nodes.Leja1D):
@@ -59,7 +59,7 @@ def evaluate_univariate_polynomial(node_gen_uni: nodes.Generator, degree: int, x
         raise TypeError(f"Unsupported node generator type: {type(node_gen_uni)}")
 
 
-def differentiate_univariate_polynomial(node_gen_uni: nodes.Generator, degree: int, x: np.ndarray) -> ArrayLike:
+def differentiate_univariate_polynomial(node_gen_uni: nodes.Generator1D, degree: int, x: np.ndarray) -> ArrayLike:
     x = node_gen_uni.scale_back(x)
     coefficients = [0] * degree + [1]
     if isinstance(node_gen_uni, nodes.Leja1D):

[…]
