### 2025-06-16 7ce830d
Attempt to use lax.scan instead of vmap
