### 2025-06-16 7cbc4e7
Make docstring more informative
