### 2025-06-16 03e84ae
Run tests with Leja nodes also on asymmetric domains
