### 2025-06-16 dc57396
Reduce test runtime (#48)
* Removed benchmark testing from CI

* Reduced number of runs within test_smolyak.py

* Removed leftover increase of recursion limit (no longer necessary for stack-based approach)

* Set default batchsize to None, which skips warm up call
