### 2025-06-16 c5170c2
Make interfaces and docstrings of nodes module clearer (#50)
* Clearer separation between 1D node sequence generators and container class, along with improved documentation

* Run tests with Leja nodes also on asymmetric domains
