+++ src/smolyax/nodes/__init__.py
@@ -1,5 +1,5 @@
-from .base import Generator
+from .base import Generator, Generator1D
 from .gausshermite import GaussHermite, GaussHermite1D
 from .leja import Leja, Leja1D
 
-__all__ = ["Generator", "Leja", "Leja1D", "GaussHermite1D", "GaussHermite"]
+__all__ = ["Generator1D", "Generator", "Leja1D", "Leja", "GaussHermite1D", "GaussHermite"]
