### 2025-06-16 b335a8f
Attempt at associative_scan
