### 2025-06-16 7a3d275
Attempt at associative_scan for gradients
