### 2025-06-16 a095eec
Set gradient to NaN if x coincides with node
