### 2025-06-16 0037e90
Implement Jacobians via analytical formula (#49)
* Remove distinction between centered and non-centered domains in evaluation of barycentric basis since it no longer offers consistent performance benefit

* Implement jacobian via analytic formula instead of autodiff

* Set gradient to NaN if x coincides with node
