### 2025-06-16 1267943
Fixed wrong dimension in runtime test
