### 2025-06-17 2fa69bc
Updated .bib
