+++ paper/paper.bib
@@ -1,4 +1,4 @@
-@book {abramowitz:1964,
+@book{abramowitz:1964,
     AUTHOR = {Abramowitz, Milton and Stegun, Irene A.},
      TITLE = {Handbook of mathematical functions with formulas, graphs, and mathematical tables},
     SERIES = {National Bureau of Standards Applied Mathematics Series},
@@ -40,7 +40,7 @@ MRREVIEWER = {Timothy\ S.\ Norfolk},
        URL = {https://doi.org/10.1023/A:1018977404843},
 }
 
-@article {bungartz:2004,
+@article{bungartz:2004,
     AUTHOR = {Bungartz, Hans-Joachim and Griebel, Michael},
      TITLE = {Sparse grids},
    JOURNAL = {Acta Numer.},
@@ -72,10 +72,9 @@ MRREVIEWER = {Mao\ Dong\ Ye},
        URL = {https://doi.org/10.1137/S0036144502417715},
 }
 
-@article {chkifa:2013,
+@article{chkifa:2013,
     AUTHOR = {Chkifa, Moulay Abdellah},
-     TITLE = {On the {L}ebesgue constant of {L}eja sequences for the complex
-              unit disk and of their real projection},
+     TITLE = {On the {L}ebesgue constant of {L}eja sequences for the complex unit disk and of their real projection},
    JOURNAL = {J. Approx. Theory},
   FJOURNAL = {Journal of Approximation Theory},
     VOLUME = {166},
@@ -89,13 +88,11 @@ MRREVIEWER = {V.\ A.\ Tkachenko},
        URL = {https://doi.org/10.1016/j.jat.2012.11.005},
 }
 
-@article {chkifa:2015,
+@article{chkifa:2015,
     AUTHOR = {Chkifa, Abdellah and Cohen, Albert and Schwab, Christoph},
-     TITLE = {Breaking the curse of dimensionality in sparse polynomial
-              approximation of parametric {PDE}s},
+     TITLE = {Breaking the curse of dimensionality in sparse polynomial approximation of parametric {PDE}s},
    JOURNAL = {J. Math. Pures Appl. (9)},
-  FJOURNAL = {Journal de Math\'ematiques Pures et Appliqu\'ees. Neuvi\`eme
-              S\'erie},
+  FJOURNAL = {Journal de Math\'ematiques Pures et Appliqu\'ees. Neuvi\`eme S\'erie},
     VOLUME = {103},
       YEAR = {2015},
     NUMBER = {2},
@@ -122,7 +119,8 @@ MRREVIEWER = {Brian\ Bradie},
     YEAR = {2018},
     NOTE = {Dissertation 25683, ETH Z\"urich},
  MRCLASS = {Thesis},
-     URL = {http://dx.doi.org/10.3929/ethz-b-000340651}
+     URL = {http://dx.doi.org/10.3929/ethz-b-000340651},
+     DOI = {10.3929/ethz-b-000340651}
 }
 
 @book{adcock:2022,
@@ -130,8 +128,7 @@ MRREVIEWER = {Brian\ Bradie},
      TITLE = {Sparse polynomial approximation of high-dimensional functions},
     SERIES = {Computational Science \& Engineering},
     VOLUME = {25},
- PUBLISHER = {Society for Industrial and Applied Mathematics (SIAM),
-              Philadelphia, PA},
+ PUBLISHER = {Society for Industrial and Applied Mathematics (SIAM), Philadelphia, PA},
       YEAR = {2022},
      PAGES = {xvii+292},
       ISBN = {978-1-611976-87-8},
@@ -141,10 +138,9 @@ MRREVIEWER = {Brian\ Bradie},
        URL = {https://doi.org/10.1137/1.9781611976885},
 }
 
-@book {dung:2023,
+@book{dung:2023,
[…]
