### 2025-06-23 e0c8b52
Moved multiplication with Smolyak coefficient into tensor product methods
