### 2025-06-23 a49cbf3
Added batching for evaluating the Smolyak interpolant and computing its gradient
