### 2025-06-26 #8204 @mjcarley
Happy to review.
