### 2025-06-26 eb5bb08
Introduce batching to avoid OOM (#51)
* Moved multiplication with Smolyak coefficient into tensor product methods

* Added batching for evaluating the Smolyak interpolant and computing its gradient
