+++ src/smolyax/interpolation.py
@@ -87,7 +87,6 @@ class SmolyakBarycentricInterpolator:
         #   - self.n_f_evals tracks the total number of function evaluations used by the interpolator
         #   - self.n_f_evals_new counts only new function calls.
         # If evaluations are reused across interpolator instances, then likely self.n_f_evals_new < self.n_f_evals
-
         self.n_f_evals = indices.nodeset_cardinality(k, t, nested=self.__is_nested)
         self.n_f_evals_new = 0
 
@@ -248,10 +247,11 @@ class SmolyakBarycentricInterpolator:
                 w_list = args[n : 2 * n]
                 s_list = args[2 * n]
                 nu = args[2 * n + 1]
-                return barycentric.evaluate_tensor_product_interpolant(x, F, xi_list, w_list, s_list, nu)
+                zeta = args[2 * n + 2]
+                return barycentric.evaluate_tensor_product_interpolant(x, F, xi_list, w_list, s_list, nu, zeta)
 
             return jax.vmap(
-                jax.jit(__evaluate_tensor_product_interpolant_wrapped), in_axes=(None, 0) + (0,) * (2 * n) + (0, 0)
+                jax.jit(__evaluate_tensor_product_interpolant_wrapped), in_axes=(None, 0) + (0,) * (2 * n + 3)
             )
 
         def __create_evaluate_tensor_product_gradient(n: int):
@@ -260,11 +260,10 @@ class SmolyakBarycentricInterpolator:
                 w_list = args[n : 2 * n]
                 s_list = args[2 * n]
                 nu = args[2 * n + 1]
-                return barycentric.evaluate_tensor_product_gradient(x, F, xi_list, w_list, s_list, nu)
+                zeta = args[2 * n + 2]
+                return barycentric.evaluate_tensor_product_gradient(x, F, xi_list, w_list, s_list, nu, zeta)
 
-            return jax.vmap(
-                jax.jit(__evaluate_tensor_product_gradient_wrapped), in_axes=(None, 0) + (0,) * (2 * n) + (0, 0)
-            )
+            return jax.vmap(jax.jit(__evaluate_tensor_product_gradient_wrapped), in_axes=(None, 0) + (0,) * (2 * n + 3))
 
         for n in self.n_2_F.keys():
             self.__compiled_tensor_product_evaluation[n] = __create_evaluate_tensor_product_interpolant(n)
@@ -275,7 +274,7 @@ class SmolyakBarycentricInterpolator:
             _ = self(inputs)
             _ = self.gradient(inputs)
 
-    def __call__(self, x: Union[jax.Array, np.ndarray]) -> jax.Array:
+    def __call__(self, x: Union[jax.Array, np.ndarray], memory_max_GB: float = 4.0) -> jax.Array:
         """@public
         Evaluate the Smolyak operator at points `x`.
 
@@ -285,6 +284,9 @@ class SmolyakBarycentricInterpolator:
             Points at which to evaluate the Smolyak interpolant of the target function `f`.
             Shape: `(n_points, d_in)` or `(d_in,)`, where `n_points` is the number of evaluation points
             and `d_in` is the dimension of the input domain.
+        memory_max_GB : float, optional
+            Maximum memory in gigabytes to use during batched evaluation. Controls the batch size
+            to stay within this memory limit. Default is 4.0.
 
         Returns
         -------
@@ -298,27 +300,45 @@ class SmolyakBarycentricInterpolator:
         if x.shape == (self.__d_in,):
             x = x[None, :]
         assert x.shape[1] == self.__d_in, f"{x.shape[1]} != {self.__d_in}"
+
         I_Lambda_x = jnp.broadcast_to(self.offset, (x.shape[0], self.__d_out))
+
[…]
