### 2025-07-06 #8505 @MoraruMaxim
@JoWestermann In the paper, you mention that other libraries exist for high-dimensional interpolation. Could you please cite some examples of related work or tools to provide context and facilitate comparison?
