### 2025-07-06 34ffe12
Fixed broken links in README (closes #52)
