### 2025-07-06 c1caa15
Setup github discussions and added hint in README (closes #54)
