### 2025-07-06 bf88547
Added info on installing with GPU support and with dev dependencies to README; added version requirements to pyproject.toml (closes #53)
