### 2025-07-06 e851e98
Fixed typo in pyproject.toml
