+++ pyproject.toml
@@ -23,7 +23,7 @@ dev = [
     "pytest-benchmark >= 5.1.0",
     "pytype >= 2024.10.11",
     "black >= 25.1.0",
-    "flake8 >= "7.3.0,
+    "flake8 >= 7.3.0",
     "flake8-pyproject >= 1.2.3",
     "pdoc >= 15.0.4",
     "mypy >= 1.16.1",
