### 2025-07-07 0f1920a
Added cuda and notebook dependency groups to pyproject.toml
