+++ pyproject.toml
@@ -18,6 +18,13 @@ dependencies = [
 requires-python = ">=3.11"
 
 [project.optional-dependencies]
+cuda = [
+    "jax[cuda] >= 0.6.2"
+]
+notebook = [
+    "ipykernel >= 6.29.5",
+    "matplotlib >= 3.10.3"
+]
 dev = [
     "pytest >= 8.4.1",
     "pytest-benchmark >= 5.1.0",
