### 2025-07-07 8982bdd
Updated installation instructions in README
