### 2025-07-08 bac0363
Fix link to docs in README (closes #58)
