+++ README.md
@@ -25,7 +25,7 @@ in well less under $0.1$ seconds on a contemporary laptop CPU.
 ### Documentation
 
 - For an introduction to relevant literature and the key implementation concepts, see the [JOSS paper (under review)](https://github.com/JoWestermann/smolyax/blob/main/paper/paper.md) accompanying this repository.
-- For code documentation see [here](https://github.com/JoWestermann/smolyax/tree/main/docs).
+- For code documentation see [here](https://jowestermann.github.io/smolyax/smolyax.html).
 
 ## Get started
 
