### 2025-07-08 e26efe3
Update action versions in CI and add pip cache
