### 2025-07-08 1c1de04
Added some references on related work
