+++ paper/paper.bib
@@ -88,6 +88,14 @@ MRREVIEWER = {V.\ A.\ Tkachenko},
        URL = {https://doi.org/10.1016/j.jat.2012.11.005},
 }
 
+@incollection{marelli:2014,
+  title={UQLab: A framework for uncertainty quantification in Matlab},
+  author={Marelli, Stefano and Sudret, Bruno},
+  booktitle={Vulnerability, uncertainty, and risk: quantification, mitigation, and management},
+  pages={2554--2563},
+  year={2014}
+}
+
 @article{chkifa:2015,
     AUTHOR = {Chkifa, Abdellah and Cohen, Albert and Schwab, Christoph},
      TITLE = {Breaking the curse of dimensionality in sparse polynomial approximation of parametric {PDE}s},
@@ -123,6 +131,19 @@ MRREVIEWER = {Brian\ Bradie},
      DOI = {10.3929/ethz-b-000340651}
 }
 
+@article{parno:2021,
+  doi = {10.21105/joss.03076},
+  url = {https://doi.org/10.21105/joss.03076},
+  year = {2021},
+  publisher = {The Open Journal},
+  volume = {6},
+  number = {68},
+  pages = {3076},
+  author = {Matthew Parno and Andrew Davis and Linus Seelinger},
+  title = {MUQ: The MIT Uncertainty Quantification Library},
+  journal = {Journal of Open Source Software}
+}
+
 @book{adcock:2022,
     AUTHOR = {Adcock, Ben and Brugiapaglia, Simone and Webster, Clayton G.},
      TITLE = {Sparse polynomial approximation of high-dimensional functions},
@@ -153,6 +174,16 @@ MRREVIEWER = {Brian\ Bradie},
        URL = {https://doi.org/10.1007/978-3-031-38384-7},
 }
 
+@article{jakeman:2023,
+ title = {PyApprox: A software package for sensitivity analysis, Bayesian inference, optimal experimental design, and multi-fidelity uncertainty quantification and surrogate modeling},
+ journal = {Environmental Modelling \& Software},
+ volume = {170},
+ pages = {105825},
+ year = {2023},
+ url = {https://doi.org/10.1016/j.envsoft.2023.105825},
+ author = {J.D. Jakeman}
+}
+
 @article{herrmann:2024,
     AUTHOR = {Herrmann, Lukas and Schwab, Christoph and Zech, Jakob},
      TITLE = {Neural and spectral operator surrogates: unified construction and expression rate bounds},
@@ -167,6 +198,16 @@ MRREVIEWER = {Brian\ Bradie},
   MRNUMBER = {4773234},
        DOI = {10.1007/s10444-024-10171-2},
        URL = {https://doi.org/10.1007/s10444-024-10171-2},
+}
+
+@article{piazzola:2024,
+ author = {Piazzola, C. and Tamellini, L.},
+ title  = {{Algorithm 1040: The Sparse Grids Matlab Kit - a Matlab implementation of sparse grids for high-dimensional function approximation and uncertainty quantification}},
+ journal= {ACM Transactions on Mathematical Software},
+ year   = {2024},
+ volume = {50},
+ number = {1},
+ doi = {10.1145/3630023}
 }
 
 @unpublished{westermann:2025,
