### 2025-07-09 44fbd40
Some more references
