+++ paper/paper.bib
@@ -113,6 +113,17 @@ MRREVIEWER = {Brian\ Bradie},
        URL = {https://doi.org/10.1016/j.matpur.2014.04.009},
 }
 
+@article{feinberg:2015,
+  author  = {Feinberg, Jonathan and Langtangen, Hans Petter},
+  title   = {Chaospy: An Open Source Tool for Designing Methods of Uncertainty Quantification},
+  journal = {Journal of Computational Science},
+  volume  = {11},
+  pages   = {46--57},
+  year    = {2015},
+  doi     = {10.1016/j.jocs.2015.08.008},
+  url     = {https://doi.org/10.1016/j.jocs.2015.08.008}
+}
+
 @software{jax:2018,
   author = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and George Necula and Adam Paszke and Jake Vander{P}las and Skye Wanderman-{M}ilne and Qiao Zhang},
   title = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
@@ -182,6 +193,16 @@ MRREVIEWER = {Brian\ Bradie},
  year = {2023},
  url = {https://doi.org/10.1016/j.envsoft.2023.105825},
  author = {J.D. Jakeman}
+}
+
+@article{tate:2023,
+  title={UncertainSCI: A Python Package for Noninvasive Parametric Uncertainty Quantification of Simulation Pipelines},
+  author={Tate, Jess and Liu, Zexin and Bergquist, Jake A and Rampersad, Sumientra and White, Dan and Charlebois, Chantel and Rupp, Lindsay and Brooks, Dana H and MacLeod, Rob S and Narayan, Akil},
+  journal={Journal of Open Source Software},
+  volume={8},
+  number={90},
+  pages={4249},
+  year={2023}
 }
 
 @article{herrmann:2024,
