### 2025-07-09 2508407
Some more references
