### 2025-07-10 9e42649
Multivariate node generators carry only copies of single 1d node generator in default case for less memory load
