### 2025-07-10 c316783
Moved datastructure setups from constructor to set_f, which allows to free memory not used later on (such as n_2_nu)
