+++ src/smolyax/interpolation.py
@@ -37,13 +37,23 @@ class SmolyakBarycentricInterpolator:
         """Output dimension of target function and interpolant"""
         return self.__d_out
 
+    @property
+    def n_f_evals(self) -> int:
+        """Number of function evaluations (== number of interpolation nodes) used by the interpolator"""
+        return self.__n_f_evals
+
+    @property
+    def n_f_evals_new(self) -> int:
+        """Number of function evaluations that were not reused from previous computations"""
+        return self.__n_f_evals_new
+
     def __init__(
         self,
         *,
+        d_out: int,
         node_gen: nodes.Generator,
         k: Sequence[float],
         t: float,
-        d_out: int,
         f: Callable[[Union[jax.Array, np.ndarray]], Union[jax.Array, np.ndarray]] = None,
         batchsize: int = None,
     ) -> None:
@@ -70,103 +80,29 @@ class SmolyakBarycentricInterpolator:
         """
         self.__d_in = len(k)
         self.__d_out = d_out
-        self.__is_nested = node_gen.is_nested
+
         self.__node_gen = node_gen
+        self.__is_nested = node_gen.is_nested
 
-        # Step 1 : Bin multiindices and smolyak coefficients by the number of active dimensions, n
-        self.__init_indices_data(k, t)
+        self.__k = k
+        self.__t = t
 
-        # Step 2 : Compute sorted dimensions and sorted degrees for all multi-indices
-        self.__init_indices_sorting()
-
-        # Step 3 : Allocate and prefill data
-        self.__init_nodes_and_weights()
-
-        # Caching the interpolation node for nu = (0,0,...,0) for reuse in self.set_f
-        self.__zero = np.array([g(0)[0] for g in self.__node_gen])
+        self.__offset = 0
+        self.__n_2_F = {}
+        self.__n_2_nodes = {}
+        self.__n_2_weights = {}
+        self.__n_2_sorted_dims = {}
+        self.__n_2_sorted_degs = {}
+        self.__n_2_zetas = {}
 
         self.__compiled_tensor_product_evaluation = {}
         self.__compiled_tensor_product_gradient = {}
 
+        self.__n_f_evals = indices.nodeset_cardinality(k, t, nested=self.__is_nested)
+        self.__n_f_evals_new = 0
+
         if f is not None:
             self.set_f(f=f, batchsize=batchsize)
-
-    def __init_indices_data(self, k: Sequence[float], t: float):
-        self.n_2_nus, self.n_2_zetas = indices.non_zero_indices_and_zetas(k, t)
-
-        # Tracking number of evaluations of the interpolation target f.
-        #   - self.n_f_evals tracks the total number of function evaluations used by the interpolator
-        #   - self.n_f_evals_new counts only new function calls.
-        # If evaluations are reused across interpolator instances, then likely self.n_f_evals_new < self.n_f_evals
-        self.n_f_evals = indices.nodeset_cardinality(k, t, nested=self.__is_nested)
-        self.n_f_evals_new = 0
-
-    def __init_indices_sorting(self):
-        self.n_2_dims = {}
-        self.n_2_sorted_dims = {}
-        self.n_2_sorted_degs = {}
-        self.n_2_argsort_dims = {}
-
-        for n in self.n_2_nus.keys():
-            if n == 0:
-                continue
-            nn = len(self.n_2_nus[n])
-            self.n_2_dims[n] = np.empty((nn, n), dtype=int)
-            self.n_2_sorted_dims[n] = np.empty((nn, n), dtype=int)
-            self.n_2_sorted_degs[n] = np.empty((nn, n), dtype=int)
-            self.n_2_argsort_dims[n] = np.empty((nn, n), dtype=int)
-
-            for i, nu in enumerate(self.n_2_nus[n]):
[…]
