### 2025-07-10 4e38c68
Merge for-loops in set_f
