+++ src/smolyax/interpolation.py
@@ -134,56 +134,49 @@ class SmolyakBarycentricInterpolator:
         if f_evals is None:
             f_evals = {}
 
-        # -------------------------------------------------------------------------------------------------------------
-        # Step 1 : Construct multiindices and smolyak coefficients, binned by the number of active dimensions n
-        # -------------------------------------------------------------------------------------------------------------
+        # Caching the interpolation node for nu = (0,0,...,0) for reuse in self.set_f
+        zero = np.array([g(0)[0] for g in self.__node_gen])
+
+        # get multi-indices and smolyak coefficients (only non-zero), binned by the number of active dimensions n
         n_2_nus, self.__n_2_zetas = indices.non_zero_indices_and_zetas(self.__k, self.__t)
-        # -------------------------------------------------------------------------------------------------------------
 
-        # -------------------------------------------------------------------------------------------------------------
-        # Step 2 : Compute sorted dimensions and sorted degrees for all multi-indices
-        # -------------------------------------------------------------------------------------------------------------
-        n_2_dims = {}
-        n_2_argsort_dims = {}
-
-        for n in n_2_nus.keys():
-            if n == 0:
-                continue
-            nn = len(n_2_nus[n])
-            n_2_dims[n] = np.empty((nn, n), dtype=int)
-            n_2_argsort_dims[n] = np.empty((nn, n), dtype=int)
-            self.__n_2_sorted_dims[n] = np.empty((nn, n), dtype=int)
-            self.__n_2_sorted_degs[n] = np.empty((nn, n), dtype=int)
-
-            for i, nu in enumerate(n_2_nus[n]):
-                n_2_dims[n][i] = list(k for k, _ in nu)
-                sorted_nu = sorted(nu, key=lambda x: x[1], reverse=True)
-                self.__n_2_sorted_dims[n][i], self.__n_2_sorted_degs[n][i] = zip(*sorted_nu)
-                n_2_argsort_dims[n][i] = np.argsort(self.__n_2_sorted_dims[n][i])
-
-        # -------------------------------------------------------------------------------------------------------------
-        # Step 3 : Allocate and prefill data
-        # -------------------------------------------------------------------------------------------------------------
         for n, nus in n_2_nus.items():
+
+            # ----- Store constant term (offset) of the Smolyak operator -----
+
             if n == 0:
-                # Smolyak constant term
                 assert len(self.__n_2_zetas[n]) == 1
-                self.__offset = self.__n_2_zetas[n][0]
+
+                nu = ()
+                if self.__is_nested:
+                    f_evals_nu = f_evals
+                else:
+                    f_evals_nu = f_evals.get(nu, {})
+                if nu not in f_evals_nu.keys():
+                    f_evals_nu[nu] = f(zero.copy())
+                    self.__n_f_evals_new += 1
+                if not self.__is_nested:
+                    f_evals[nu] = f_evals_nu
+
+                self.__offset = self.__n_2_zetas[n][0] * f_evals_nu[nu]
                 continue
+
+            # ----- Assemble sorted dimensions and sorted degrees -----
 
             nn = len(nus)  # number of multi-indices of length n
-            sorted_degs = self.__n_2_sorted_degs[n]
-            sorted_dims = self.__n_2_sorted_dims[n]
+            sorted_dims = np.empty((nn, n), dtype=int)
+            sorted_degs = np.empty((nn, n), dtype=int)
+
[…]
