### 2025-07-10 034013d
Merge for-loops in set_f
