### 2025-07-11 783ae03
Moved memory limit arg from __call__ to constructor and renamed batchsize to n_inputs
