+++ src/smolyax/interpolation.py
@@ -55,7 +55,8 @@ class SmolyakBarycentricInterpolator:
         k: Sequence[float],
         t: float,
         f: Callable[[Union[jax.Array, np.ndarray]], Union[jax.Array, np.ndarray]] = None,
-        batchsize: int = None,
+        n_inputs: int = None,
+        memory_limit: float = 4.0,
     ) -> None:
         r"""
         Initialize the Smolyak Barycentric Interpolator.
@@ -74,9 +75,12 @@ class SmolyakBarycentricInterpolator:
             Target function to interpolate. While `f` can be passed at construction time, for a better control over, and
             potential reuse of, function evaluations consider calling [`set_f()`](#SmolyakBarycentricInterpolator.set_f)
             *after* construction.
-        batchsize : int, default=None
-            Expected batch size of the interpolator input, used to pre-compile the `__call__` method via a warm-up call.
-            If `batchsize` is `None`, the warm-up is skipped.
+        n_inputs : int, default=None
+            Expected number of input samples, used to pre-compile the `__call__` method via a warm-up call.
+            If `n_input` is `None`, the warm-up is skipped and `n_input` is set upon first use of `__call__`.
+        memory_limit : float, optional
+            Maximum memory in gigabytes to use during batched evaluation. Controls the batch size
+            to stay within this memory limit. Default is 4.0.
         """
         self.__d_in = len(k)
         self.__d_out = d_out
@@ -95,21 +99,23 @@ class SmolyakBarycentricInterpolator:
         self.__n_2_sorted_degs = {}
         self.__n_2_zetas = {}
 
-        self.__compiled_tensor_product_evaluation = {}
-        self.__compiled_tensor_product_gradient = {}
-
         self.__n_f_evals = indices.nodeset_cardinality(k, t, nested=self.__is_nested)
         self.__n_f_evals_new = 0
 
+        self.__compiled_tensor_product_evaluation = {}
+        self.__compiled_tensor_product_gradient = {}
+
+        self.__memory_limit = memory_limit
+        self.__n_inputs = n_inputs
+
         if f is not None:
-            self.set_f(f=f, batchsize=batchsize)
+            self.set_f(f=f)
 
     def set_f(
         self,
         *,
         f: Callable[[Union[jax.Array, np.ndarray]], Union[jax.Array, np.ndarray]],
         f_evals: dict[tuple, dict[tuple, jax.Array]] = None,
-        batchsize: int = None,
     ) -> dict[tuple, dict[tuple, jax.Array]]:
         """
         Compute (or reuse pre-computed) evaluations of the target function `f` at the interpolation nodes of the
@@ -122,9 +128,6 @@ class SmolyakBarycentricInterpolator:
         f_evals : dict, optional
             A dictionary mapping interpolation nodes to function evaluations.
             If provided, these evaluations will be reused.
-        batchsize : int, default=None
-            Expected batch size of the interpolator input, used to pre-compile the `__call__` method via a warm-up call.
-            If `batchsize` is `None`, the warm-up is skipped.
 
         Returns
         -------
@@ -230,11 +233,11 @@ class SmolyakBarycentricInterpolator:
             self.__n_2_weights[n] = [jnp.array(w) for w in weights]
             self.__n_2_zetas[n] = jnp.array(self.__n_2_zetas[n])
 
-        self.__compile_for_batchsize(batchsize)
+        self.__setup_eval_functions()
 
         return f_evals
 
-    def __compile_for_batchsize(self, batchsize: int) -> None:
+    def __setup_eval_functions(self) -> None:
 
         def __create_evaluate_tensor_product_interpolant(n: int):
[…]
