### 2025-07-11 fd731d7
Added method __validate_input
