### 2025-07-11 027f968
Added method __validate_input
