### 2025-07-14 7701eac
No longer using python list when calling jitted function
