+++ src/smolyax/interpolation.py
@@ -166,8 +166,9 @@ class SmolyakBarycentricInterpolator:
             # ----- Assemble nodes and weights -----
 
             tau = tuple(int(ti) for ti in sorted_degs.max(axis=0))  # per-dimension maximal degree tau_i
-            nodes = [np.zeros((nn, tau_i + 1), dtype=float) for tau_i in tau]
-            weights = [np.zeros((nn, tau_i + 1), dtype=float) for tau_i in tau]
+            tau_max = max(tau)
+            nodes = np.zeros((nn, n, tau_max + 1), dtype=float)
+            weights = np.zeros((nn, n, tau_max + 1), dtype=float)
 
             # for each slot t, group i's by (dim,deg) so we only gen once
             for t in range(n):
@@ -181,9 +182,8 @@ class SmolyakBarycentricInterpolator:
                 for (dim, deg), idxs in groups.items():
                     pts = self.__node_gen[dim](deg)
                     wts = barycentric.compute_weights(pts)
-                    L = len(pts)
-                    nodes[t][idxs, :L] = pts
-                    weights[t][idxs, :L] = wts
+                    nodes[idxs, t, : deg + 1] = pts
+                    weights[idxs, t, : deg + 1] = wts
                 # we can do even better if we vectorize node_gen(degrees) for isotropic rules, like GH or Leja
 
             # ----- Assemble the array storing the functions evaluations -----
@@ -206,7 +206,7 @@ class SmolyakBarycentricInterpolator:
                 for mu_degrees in it.product(*ranges):
                     mu_tuple = tuple((s_i[i], mu_degrees[i]) for i in argsort_s_i if mu_degrees[i] > 0)
                     if mu_tuple not in f_evals_nu:
-                        x[s_i] = [xi_k[i][deg] for xi_k, deg in zip(nodes, mu_degrees)]
+                        x[s_i] = [xi_k[deg] for xi_k, deg in zip(nodes[i], mu_degrees)]
                         f_evals_nu[mu_tuple] = f(x)
                         self.__n_f_evals_new += 1
                     F_i[:, *mu_degrees] = f_evals_nu[mu_tuple]
@@ -218,8 +218,8 @@ class SmolyakBarycentricInterpolator:
             self.__n_2_F[n] = jnp.array(F)
             self.__n_2_sorted_degs[n] = jnp.array(sorted_degs)
             self.__n_2_sorted_dims[n] = jnp.array(sorted_dims)
-            self.__n_2_nodes[n] = [jnp.array(xi) for xi in nodes]
-            self.__n_2_weights[n] = [jnp.array(w) for w in weights]
+            self.__n_2_nodes[n] = jnp.array(nodes)
+            self.__n_2_weights[n] = jnp.array(weights)
             self.__n_2_zetas[n] = jnp.array(self.__n_2_zetas[n])
 
         self.__setup_eval_functions()
@@ -228,33 +228,13 @@ class SmolyakBarycentricInterpolator:
 
     def __setup_eval_functions(self) -> None:
 
-        def __create_evaluate_tensor_product_interpolant(n: int):
-            def __evaluate_tensor_product_interpolant_wrapped(x, F, *args):
-                xi_list = args[:n]
-                w_list = args[n : 2 * n]
-                s_list = args[2 * n]
-                nu = args[2 * n + 1]
-                zeta = args[2 * n + 2]
-                return barycentric.evaluate_tensor_product_interpolant(x, F, xi_list, w_list, s_list, nu, zeta)
-
-            return jax.vmap(
-                jax.jit(__evaluate_tensor_product_interpolant_wrapped), in_axes=(None, 0) + (0,) * (2 * n + 3)
-            )
-
-        def __create_evaluate_tensor_product_gradient(n: int):
[…]
