### 2025-07-14 50956a7
Fixed missing slicing in __call__ loop
