### 2025-07-14 f236609
Removed n-specific function handles
