### 2025-07-14 3a1a695
Fixed loop index n
