### 2025-07-15 ee16303
Update documentation (#61)
* Move from checked in docs to github deployed docs

* Update action versions in CI and add pip cache

* Removed pdoc call from README

* Add landing page for documentation
