### 2025-07-15 b09fba5
Use dependency groups for clearer installation instructions (#57)
* Added cuda and notebook dependency groups to pyproject.toml

* Updated installation instructions in README
