### 2025-07-15 4bbea0e
Fix build error by installing project and dependencies (#61)
