### 2025-07-21 #8505 @mjcarley
I am mostly happy with the software and documentation, but I would like to see some comparison to other methods and/or packages. A search for "polynomial interpolation" on github throws up many results, but these seem to be of varying quality or completeness. Would it be possible to give a comparison to one or two roughly equivalent packages on ACM TOMS or elsewhere?
