### 2025-07-21 #8505 @JoWestermann
Dear reviewers, thank you very much for your comments here and over in the repo. We plan to address the request for citation of and comparison to related work with JoWestermann/smolyax/pull/60. We already included a number of references and are working on a benchmark comparison with one other code base.

I will be traveling 07/26 until 08/02, so I hope to wrap this up before that week but in any case I will post an update here.
