+++ .github/workflows/ci.yml
@@ -10,9 +10,10 @@ jobs:
         python-version: ["3.9", "3.12"] # Define the Python versions to test against
     steps:
       - uses: actions/checkout@v4
-      - name: Set up Python
+      - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v5
         with:
+          python-version: ${{ matrix.python-version }}
           cache: pip
       - name: Install dependencies
         run: pip install .[dev]
