### 2025-07-22 bee80db
Update ci.yml (#64)
* Update ci.yml

* Update ci.yml
