### 2025-07-23 8c6dbd2
No longer using python lists to pass nodes and weigths to evaluate_tensor_product
