### 2025-07-23 5a84b97
No longer using python lists to pass nodes and weigths to evaluate_tensor_product
