### 2025-07-24 00457a5
Revert "Update ci.yml (#64)" -- wait for the batching branch/pr to be merged, then add this via the python39 branch/pr
This reverts commit d7df0f6dbb210de6fa5a8ccf2ffa0c5f16510089.
