### 2025-07-26 00cdac7
Scratchwork, fast output dimension
