+++ src/smolyax/interpolation.py
@@ -199,7 +199,7 @@ class SmolyakBarycentricInterpolator:
 
             # ----- Assemble the array storing the functions evaluations -----
 
-            F = np.zeros((nn, self.__d_out) + tuple(ti + 1 for ti in tau), dtype=float)
+            F = np.zeros((nn,) + tuple(ti + 1 for ti in tau) + (self.__d_out,), dtype=float)
 
             for i, nu in enumerate(n_2_nus[n]):
                 x = zero.copy()
@@ -214,19 +214,20 @@ class SmolyakBarycentricInterpolator:
                 argsort_s_i = np.argsort(s_i)
 
                 ranges = [range(k + 1) for k in sorted_degs[i]]
+
                 for mu_degrees in it.product(*ranges):
                     mu_tuple = tuple((s_i[i], mu_degrees[i]) for i in argsort_s_i if mu_degrees[i] > 0)
                     if mu_tuple not in f_evals_nu:
                         x[s_i] = [xi_k[i][deg] for xi_k, deg in zip(nodes, mu_degrees)]
                         f_evals_nu[mu_tuple] = f(x)
                         self.__n_f_evals_new += 1
-                    F_i[:, *mu_degrees] = f_evals_nu[mu_tuple]
+                    F_i[mu_degrees] = f_evals_nu[mu_tuple]
 
                 if not self.__is_nested:
                     f_evals[nu] = f_evals_nu
 
             # save as jnp data structures
-            self.__n_2_F[n] = jnp.array(F)
+            self.__n_2_F[n] = jnp.array(F) #n nodes with such dims, dim_1, dim_2, ...dim_3, out_dim
             self.__n_2_sorted_degs[n] = jnp.array(sorted_degs)
             self.__n_2_sorted_dims[n] = jnp.array(sorted_dims)
             self.__n_2_nodes[n] = [jnp.array(xi) for xi in nodes]
@@ -238,39 +239,41 @@ class SmolyakBarycentricInterpolator:
         return f_evals
 
     def __setup_eval_functions(self) -> None:
-
+        # if self.__n_inputs < 500:
+        #     evaluator = barycentric.evaluate_tensor_product_interpolant_low_N
+        # else:
+        evaluator = barycentric.evaluate_tensor_product_interpolant
         def __create_evaluate_tensor_product_interpolant(n: int):
             def __evaluate_tensor_product_interpolant_wrapped(x, F, *args):
                 xi_list = args[:n]
                 w_list = args[n : 2 * n]
                 s_list = args[2 * n]
-                nu = args[2 * n + 1]
-                zeta = args[2 * n + 2]
-                return barycentric.evaluate_tensor_product_interpolant(x, F, xi_list, w_list, s_list, nu, zeta)
-
+                zeta = args[2 * n + 1]
+                return evaluator(x, F, xi_list, w_list, s_list, zeta)
+ 
             return jax.vmap(
-                jax.jit(__evaluate_tensor_product_interpolant_wrapped), in_axes=(None, 0) + (0,) * (2 * n + 3)
+                jax.jit(__evaluate_tensor_product_interpolant_wrapped), in_axes=(None, 0) + (0,) * (2 * n + 2)
             )
 
-        def __create_evaluate_tensor_product_gradient(n: int):
-            def __evaluate_tensor_product_gradient_wrapped(x, F, *args):
-                xi_list = args[:n]
-                w_list = args[n : 2 * n]
-                s_list = args[2 * n]
-                nu = args[2 * n + 1]
-                zeta = args[2 * n + 2]
-                return barycentric.evaluate_tensor_product_gradient(x, F, xi_list, w_list, s_list, nu, zeta)
+        # def __create_evaluate_tensor_product_gradient(n: int):
[…]
