+++ src/smolyax/barycentric.py
@@ -29,7 +29,6 @@ def compute_weights(nodes: Union[jax.Array, np.ndarray]) -> jax.Array:
     diffs = jnp.where(diffs == 0, 1, diffs)
     return jnp.prod(1 / diffs, axis=0)
 
-
 def evaluate_basis_unnormalized(x: jax.Array, xi: jax.Array, w: jax.Array, nu_i: int) -> jax.Array:
     r"""
     Evaluate the barycentric basis numerator terms at given evaluation points.
@@ -56,22 +55,71 @@ def evaluate_basis_unnormalized(x: jax.Array, xi: jax.Array, w: jax.Array, nu_i:
         Barycentric numerator terms for indices $j \le \nu_i$, or a one-hot indicator pattern if `x` coincides with a
         node. Shape `(n_points, m_i)`.
     """
-    diffs = x - xi
-    mask_cols = jnp.arange(diffs.shape[1]) <= nu_i
-    mask_zero = jnp.any((diffs == 0) & mask_cols, axis=1)
-    one_hot_pattern = (diffs == 0).astype(diffs.dtype)
-    w_div_diffs = jnp.divide(w, diffs)
-    b = jnp.where(mask_zero[:, None], one_hot_pattern, w_div_diffs)
-    return jnp.where(mask_cols[None, :], b, 0)
+    # diffs = x - xi
+    # mask_cols = jnp.arange(diffs.shape[1]) <= nu_i
+    # mask_zero = jnp.any((diffs == 0) & mask_cols, axis=1)
+    # one_hot_pattern = (diffs == 0).astype(diffs.dtype)
+    # w_div_diffs = jnp.divide(w, diffs)
+    # b = jnp.where(mask_zero[:, None], one_hot_pattern, w_div_diffs)
+    # return jnp.where(mask_cols[None, :], b, 0)
 
+    diffs  = x - xi[None, :]               # (N,μ)
+    zeros  = (diffs == 0)                  # (N,μ)
+    # any0   = jnp.any(zeros, axis=1)        # (N,)
+    w_div  = w[None, :] / diffs            # (N,μ)
+    b      = jnp.where(                    # (N,μ)
+               jnp.any(zeros, axis=1)[:,None],
+               zeros.astype(w_div.dtype),
+               w_div
+             )
+    s      = jnp.sum(b, axis=1)           # (N,)
+    return b, s
 
+@jax.jit
+def evaluate_basis_unnormalized_no_nu(x: jax.Array, xi: jax.Array, w: jax.Array) -> jax.Array:
+    r"""
+    Evaluate the barycentric basis numerator terms at given evaluation points.
+
+    Computes $w_j / (x - \xi_j)$ for indices $j \le \nu_i$; other entries are masked to zero.
+    If an evaluation point coincides with a node $\xi_j$ for $j \le \nu_i$, returns a one-hot indicator pattern
+    restricted to those entries.
+
+    Parameters
+    ----------
+    x : jax.Array
+        Evaluation points along one dimension. Shape: `(n_points, 1)`
+    xi : jax.Array
+        Interpolation nodes. Shape `(m_i,)` with `m_i > nu_i`. Entries beyond `nu_i` are not used.
+    w : jax.Array
+        Barycentric weights corresponding to the given interpolation nodes. Shape `(m_i,)` with `m_i > nu_i`.
+        Entries beyond `nu_i` are not used.
+    nu_i : int
+        Polynomial degree in this dimension.
+
+    Returns
+    -------
+    b : jax.Array
+        Barycentric numerator terms for indices $j \le \nu_i$, or a one-hot indicator pattern if `x` coincides with a
+        node. Shape `(n_points, m_i)`.
+    """
+    diffs  = x - xi[None, :]               # (N,μ)
[…]
