+++ setup.cfg
@@ -2,4 +2,4 @@
 packages = find:
 package_dir = 
     =src
-python_requires = >=3.11
+python_requires = >=3.9
