+++ pyproject.toml
@@ -12,14 +12,14 @@ authors = [
 ]
 license = { text = "MIT" }
 dependencies = [
-    "jax >= 0.6.2",
-    "numba >= 0.61.2",
+    "jax",
+    "numba",
 ]
-requires-python = ">=3.11"
+requires-python = ">=3.9"
 
 [project.optional-dependencies]
 cuda = [
-    "jax[cuda] >= 0.6.2"
+    "jax[cuda]"
 ]
 notebook = [
     "ipykernel >= 6.29.5",
@@ -28,7 +28,7 @@ notebook = [
 dev = [
     "pytest >= 8.4.1",
     "pytest-benchmark >= 5.1.0",
-    "pytype >= 2024.10.11",
+    "pytype",
     "black >= 25.1.0",
     "flake8 >= 7.3.0",
     "flake8-pyproject >= 1.2.3",
@@ -40,14 +40,14 @@ dev = [
 
 [tool.black]
 line-length = 120
-target-version = ['py311']
+target-version = ['py39']
 
 [tool.flake8]
 max-line-length = 120
 extend-ignore = """E203"""
 
 [tool.pytype]
-python_version = "3.11"
+python_version = "3.9"
 inputs = ["src/smolyax"]
 disable = ["import-error"]
 
