### 2025-08-04 4e3a292
Use single einsum call in evaluate_tensor_product_interpolant
