### 2025-08-05 88fbb41
Simplified quadrature call
