### 2025-08-05 655b3ba
Simplified quadrature call
