### 2025-08-05 3e1d4e4
Simplified einsum index generation in tensor product interpolant
