### 2025-08-06 1d7dd09
Attempt at vectorizing barycentric util methods (works for interpolation but not yet gradients)
