+++ src/smolyax/interpolation.py
@@ -93,7 +93,7 @@ class SmolyakBarycentricInterpolator:
         self.__n_f_evals_new = 0
 
         self.__compiled_tensor_product_evaluation = None
-        self.__compiled_tensor_product_gradient = None
+        self.__compiled_tensor_product_gradient = {}
 
         self.__memory_limit = memory_limit
         self.__n_inputs = n_inputs
@@ -167,9 +167,8 @@ class SmolyakBarycentricInterpolator:
             # ----- Assemble nodes and weights -----
 
             tau = tuple(int(ti) for ti in sorted_degs.max(axis=0))  # per-dimension maximal degree tau_i
-            tau_max = max(tau)
-            nodes = np.zeros((nn, n, tau_max + 1), dtype=float)
-            weights = np.zeros((nn, n, tau_max + 1), dtype=float)
+            nodes = [np.zeros((nn, tau_i + 1), dtype=float) for tau_i in tau]
+            weights = [np.zeros((nn, tau_i + 1), dtype=float) for tau_i in tau]
 
             # for each slot t, group i's by (dim,deg) so we only gen once
             for t in range(n):
@@ -183,8 +182,10 @@ class SmolyakBarycentricInterpolator:
                 for (dim, deg), idxs in groups.items():
                     pts = self.__node_gen[dim](deg)
                     wts = barycentric.compute_weights(pts)
-                    nodes[idxs, t, : deg + 1] = pts
-                    weights[idxs, t, : deg + 1] = wts
+                    L = len(pts)
+                    nodes[t][idxs, :L] = pts
+                    weights[t][idxs, :L] = wts
+
                 # we can do even better if we vectorize node_gen(degrees) for isotropic rules, like GH or Leja
 
             # ----- Assemble the array storing the functions evaluations -----
@@ -207,7 +208,7 @@ class SmolyakBarycentricInterpolator:
                 for mu_degrees in it.product(*ranges):
                     mu_tuple = tuple((s_i[i], mu_degrees[i]) for i in argsort_s_i if mu_degrees[i] > 0)
                     if mu_tuple not in f_evals_nu:
-                        x[s_i] = [xi_k[deg] for xi_k, deg in zip(nodes[i], mu_degrees)]
+                        x[s_i] = [xi_k[i][deg] for xi_k, deg in zip(nodes, mu_degrees)]
                         f_evals_nu[mu_tuple] = f(x)
                         self.__n_f_evals_new += 1
                     F_i[:, *mu_degrees] = f_evals_nu[mu_tuple]
@@ -219,8 +220,8 @@ class SmolyakBarycentricInterpolator:
             self.__n_2_F[n] = jnp.array(F)
             self.__n_2_sorted_degs[n] = jnp.array(sorted_degs)
             self.__n_2_sorted_dims[n] = jnp.array(sorted_dims)
-            self.__n_2_nodes[n] = jnp.array(nodes)
-            self.__n_2_weights[n] = jnp.array(weights)
+            self.__n_2_nodes[n] = [jnp.array(xi) for xi in nodes]
+            self.__n_2_weights[n] = [jnp.array(w) for w in weights]
             self.__n_2_zetas[n] = jnp.array(self.__n_2_zetas[n])
 
         self.__setup_eval_functions()
@@ -229,12 +230,22 @@ class SmolyakBarycentricInterpolator:
 
     def __setup_eval_functions(self) -> None:
 
-        self.__compiled_tensor_product_evaluation = jax.jit(
-            jax.vmap(barycentric.evaluate_tensor_product_interpolant, in_axes=(None, 0, 0, 0, 0, 0, 0))
-        )
-        self.__compiled_tensor_product_gradient = jax.jit(
[…]
