+++ src/smolyax/barycentric.py
@@ -31,6 +31,45 @@ def compute_weights(nodes: Union[jax.Array, np.ndarray]) -> jax.Array:
     return jnp.prod(1 / diffs, axis=0)
 
 
+def evaluate_basis_unnormalized_vectorized(x: jax.Array, xi: jax.Array, w: jax.Array, nu_i: int) -> jax.Array:
+    r"""
+    Evaluate the barycentric basis numerator terms at given evaluation points.
+
+    Computes $w_j / (x - \xi_j)$ for indices $j \le \nu_i$; other entries are masked to zero.
+    If an evaluation point coincides with a node $\xi_j$ for $j \le \nu_i$, returns a one-hot indicator pattern
+    restricted to those entries.
+
+    Parameters
+    ----------
+    x : jax.Array
+        Evaluation points along one dimension. Shape: `(n_points, 1)`
+    xi : jax.Array
+        Interpolation nodes. Shape `(m_i,)` with `m_i > nu_i`. Entries beyond `nu_i` are not used.
+    w : jax.Array
+        Barycentric weights corresponding to the given interpolation nodes. Shape `(m_i,)` with `m_i > nu_i`.
+        Entries beyond `nu_i` are not used.
+    nu_i : int
+        Polynomial degree in this dimension.
+
+    Returns
+    -------
+    b : jax.Array
+        Barycentric numerator terms for indices $j \le \nu_i$, or a one-hot indicator pattern if `x` coincides with a
+        node. Shape `(n_points, m_i)`.
+    """
+    diffs = x[:, :, None] - xi[:, None, :]
+
+    mask_cols = jnp.arange(xi.shape[-1])[None, :] <= nu_i[:, None]
+    mask_cols = mask_cols[:, None, :]
+
+    mask_zero = jnp.any((diffs == 0) & mask_cols, axis=-1)
+    one_hot = (diffs == 0).astype(diffs.dtype)
+    w_div_diff = w[:, None, :] / diffs
+
+    b = jnp.where(mask_zero[..., None], one_hot, w_div_diff)
+    return jnp.where(mask_cols, b, 0)
+
+
 def evaluate_basis_unnormalized(x: jax.Array, xi: jax.Array, w: jax.Array, nu_i: int) -> jax.Array:
     r"""
     Evaluate the barycentric basis numerator terms at given evaluation points.
@@ -66,6 +105,7 @@ def evaluate_basis_unnormalized(x: jax.Array, xi: jax.Array, w: jax.Array, nu_i:
     return jnp.where(mask_cols[None, :], b, 0)
 
 
+@jax.jit
 def evaluate_tensor_product_interpolant(
     x: jax.Array,
     F: jax.Array,
@@ -111,18 +151,20 @@ def evaluate_tensor_product_interpolant(
         The evaluated tensor product interpolant at the points specified by `x`.
         The shape of the output will be `(n_points, d_out)`.
     """
-    bs = [
-        evaluate_basis_unnormalized(x[:, [si]], ni[:i], wi[:i], nui)
-        for (i, si, nui, ni, wi) in zip(F.shape[1:], sorted_dims, sorted_degs, xi_list, w_list)
-    ]
+    x_sel = jnp.take(x, sorted_dims, axis=1).transpose(2, 1, 0)
+    bs = []
+    for i in range(sorted_degs.shape[1]):
+        b = evaluate_basis_unnormalized_vectorized(x_sel[i], xi_list[i], w_list[i], sorted_degs[:, i])
[…]
