### 2025-08-06 06ff457
Small simplifications
