### 2025-08-06 69e8246
Reworked related work paragraph
