### 2025-08-06 ace1737
Reworked related work paragraph
