+++ paper/paper.bib
@@ -124,6 +124,15 @@ MRREVIEWER = {Brian\ Bradie},
   url     = {https://doi.org/10.1016/j.jocs.2015.08.008}
 }
 
+@TechReport{tasmanian,
+  title={User Manual: TASMANIAN Sparse Grids},
+  author = "Stoyanov, M",
+  institution = "Oak Ridge National Laboratory",
+  address = "One Bethel Valley Road, Oak Ridge, TN",
+  number = "ORNL/TM-2015/596",
+  year = "2015"
+}
+
 @software{jax:2018,
   author = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and George Necula and Adam Paszke and Jake Vander{P}las and Skye Wanderman-{M}ilne and Qiao Zhang},
   title = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
