### 2025-08-06 c8f9adc
Added benchmark code
