### 2025-08-06 3193fbd
Added benchmark code
