+++ benchmarking/benchmark.py
@@ -1,0 +1,250 @@
+# ----------------------------------------------------------------------------
+# Installing TASMANIAN
+#   * export CMAKE_CUDA_COMPILER=/usr/local/cuda-12.2/bin/nvcc
+#   * export Tasmanian_ENABLE_CUDA=/usr/local/cuda-12.2/bin/nvcc
+#   * pip install --no-cache-dir Tasmanian --no-binary=Tasmanian Tasmanian
+# ----------------------------------------------------------------------------
+import argparse
+import GPUtil
+import os
+import socket
+import time
+from pathlib import Path
+
+import numpy as np
+import matplotlib.pyplot as plt
+from matplotlib.colors import TwoSlopeNorm
+
+import jax
+jax.config.update("jax_platform_name", "gpu")
+
+import Tasmanian
+from smolyax import indices, nodes
+from smolyax.interpolation import SmolyakBarycentricInterpolator
+
+import testfunction, depthmap
+
+
+# ----------------------------------------------------------------------------
+# USER PARAMETERS
+# ----------------------------------------------------------------------------
+
+
+TARGET_N_LIST = [1_000, 2_000, 4_000, 6_000, 8_000, 10_000]
+N_MC_LIST     = [50, 500, 5000]
+D_IN_LIST     = [10, 40, 160]
+D_OUT_LIST    = [10, 40, 160]
+
+N_ITER = 20
+
+SMOLYAX_MEMORY_LIMIT = 20.  # Adjust to GPU capabilities (leave a little bit space for overhead)
+
+RESULTS_DIR = Path("results")
+RESULTS_DIR.mkdir(exist_ok=True)
+
+HOST = socket.gethostname()
+FLAG = '_1'
+
+# ----------------------------------------------------------------------------
+# HELPER FUNCTIONS
+# ----------------------------------------------------------------------------
+
+
+def tasmanian_filename(d_in: int, d_out: int, depth: int, n_mc: int) -> Path:
+    return RESULTS_DIR / (
+        f"tasmanian_din{d_in}_dout{d_out}_depth{depth}_nmc{n_mc}_niter{N_ITER}_{HOST}.npz"
+    )
+
+
+def smolyax_filename(d_in: int, d_out: int, N_target: int, n_mc: int) -> Path:
+    return RESULTS_DIR / (
+        f"smolyax_din{d_in}_dout{d_out}_N{N_target}_nmc{n_mc}_niter{N_ITER}_{HOST}_{SMOLYAX_MEMORY_LIMIT}{FLAG}.npz"
+    )
+
+
+def rmse(y_true, y_pred) :
+    return np.sqrt(np.mean((y_true - y_pred) ** 2) / np.mean(y_true ** 2))
+
+
+# ----------------------------------------------------------------------------
+# BENCHMARK FUNCTIONS
+# ----------------------------------------------------------------------------
+
+
+def run_tasmanian(
+    d_in: int,
+    d_out: int,
+    depth: int,
+    n_mc: int,
+    f_test,
+    aniso,
+    X_test,
+    Y_test,
+    use_saved=True,
+):
+    fname = tasmanian_filename(d_in, d_out, depth, n_mc)
+    if fname.is_file() and use_saved:
+        d = np.load(fname)
+        return float(d["e"][()]), int(d["N"][()]), float(d["t"][()])
+
+    grid = Tasmanian.makeSequenceGrid(d_in, d_out, depth, "hyperbolic", "leja", aniso)
+    grid.enableAcceleration("gpu-cublas", 0)  # other available GPU backends do not work for d_in >= 32
+    assert grid.isCudaEnabled()
+
[…]
