### 2025-08-07 d5cf05c
Fuse einsum calls also in gradient
