### 2025-08-07 3d04a7b
Python39 cherry pick (#66)
* Python 3.9 compatability

* Update ci.yml

* Update pyproject.toml

* Updated Readme to new python requirement

---------

Co-authored-by: Joshua Chen <joshuawchen@icloud.com>
Co-authored-by: joshuawchen <joshuawchen@utexas.edu>
