+++ pyproject.toml
@@ -12,10 +12,10 @@ authors = [
 ]
 license = { text = "MIT" }
 dependencies = [
-    "jax >= 0.6.2",
-    "numba >= 0.61.2",
+    "jax >= 0.4.30",
+    "numba >= 0.60.0"
 ]
-requires-python = ">=3.11"
+requires-python = ">=3.9"
 
 [project.optional-dependencies]
 cuda = [
@@ -28,14 +28,14 @@ notebook = [
 dev = [
     "pytest >= 8.4.1",
     "pytest-benchmark >= 5.1.0",
-    "pytype >= 2024.10.11",
+    "pytype",
     "black >= 25.1.0",
     "flake8 >= 7.3.0",
     "flake8-pyproject >= 1.2.3",
     "pdoc >= 15.0.4",
     "mypy >= 1.16.1",
     "ipykernel >= 6.29.5",
-    "matplotlib >= 3.10.3"
+    "matplotlib >= 3.9.4"
 ]
 
 [tool.black]
