### 2025-08-08 #8505 @JoWestermann
Dear reviewers, thank you very much for your feedback and your patience. I believe we have now addressed all of your comments, see in particular JoWestermann/smolyax/pull/60. Please let us know if any of you have further remarks.
