### 2025-08-11 07a2dbe
Update paper.md
