+++ paper/paper.bib
@@ -93,6 +93,7 @@ MRREVIEWER = {V.\ A.\ Tkachenko},
   author={Marelli, Stefano and Sudret, Bruno},
   booktitle={Vulnerability, uncertainty, and risk: quantification, mitigation, and management},
   pages={2554--2563},
+  doi={10.1061/9780784413609.257},
   year={2014}
 }
 
@@ -211,6 +212,7 @@ MRREVIEWER = {Brian\ Bradie},
   volume={8},
   number={90},
   pages={4249},
+  doi={10.21105/joss.04249},
   year={2023}
 }
 
