### 2025-08-13 0113e5f
Removed redundant setup.cfg and pytype.cfg
