+++ pytype.cfg
@@ -1,4 +1,0 @@
-[pytype]
-python_version = 3.11
-inputs = src/smolyax
-disable = import-error
