+++ setup.cfg
@@ -1,5 +1,0 @@
-[options]
-packages = find:
-package_dir = 
-    =src
-python_requires = >=3.11
