### 2025-08-13 66a6327
Minor pyproject.toml updates
