### 2025-08-13 #8505 @JoWestermann
@diehlpk Done. Authors, affiliations etc were double checked; we now have a release for v1.0.0; and the zenodo archive has the DOI 10.5281/zenodo.16860779.
