### 2025-08-14 67db074
Update paper.md
