+++ paper/paper.md
@@ -53,9 +53,9 @@ The `smolyax` library provides interpolation capabilities for arbitrary multivar
 
 # Statement of Need
 
-Polynomial expansion is a well-studied and powerful tool in applied mathematics, with important applications in surrogate modeling, uncertainty quantification and inverse problems, see e.g. [@adcock:2022; @dung:2023; @zech:2018; @chkifa:2015; @herrmann:2024; @westermann:2025] and references therein. Smolyak interpolation offers a practical way to construct polynomial approximations with known error bounds for a wide range of function classes, see e.g. [@barthelmann:2000; @chkifa:2015; @adcock:2022].
+Polynomial expansion is a well-studied and powerful tool in applied mathematics, with important applications in surrogate modeling, uncertainty quantification and inverse problems, see e.g., @adcock:2022, @dung:2023, @zech:2018, @chkifa:2015, @herrmann:2024, @westermann:2025, and references therein. Smolyak interpolation offers a practical way to construct polynomial approximations with known error bounds for a wide range of function classes, see e.g., @barthelmann:2000, @chkifa:2015, and @adcock:2022.
 
-Several libraries provide CPU-based high-dimensional interpolation functionality, for example `Chaospy` [@feinberg:2015], `UQLab` [@marelli:2014], `The Sparse Grid Matlab Kit` [@piazzola:2024], `PyApprox` [@jakeman:2023], `MUQ` [@parno:2021] or `UncertainSCI` [@tate:2023]. The GPU support that is in practice necessary to go from moderate to high dimensions is offered so far only by `Tasmanian` [@tasmanian]. Benchmark experiments suggest that while asymptotic runtime of the Smolyak interpolator in `Tasmanian` scale better as the output dimensions $d_{\rm out}$ increases, `smolyax` offers competitive asymptotic runtimes for increasing $d_{\rm in}$ and input data set size.
+Several libraries provide CPU-based high-dimensional interpolation functionality, for example `Chaospy` [@feinberg:2015], `UQLab` [@marelli:2014], `The Sparse Grid Matlab Kit` [@piazzola:2024], `PyApprox` [@jakeman:2023], `MUQ` [@parno:2021], and `UncertainSCI` [@tate:2023]. The GPU support that is necessary in practice to go from moderate to high dimensions is offered so far only by `Tasmanian` [@tasmanian]. Benchmark experiments suggest that while asymptotic runtime of the Smolyak interpolator in `Tasmanian` scale better as the output dimensions $d_{\rm out}$ increases, `smolyax` offers competitive asymptotic runtimes for increasing $d_{\rm in}$ and input data set size.
 
 # A vectorizable implementation of the Smolyak operator
 
@@ -69,7 +69,7 @@ Smolyak interpolation [@smolyak:1963; @barthelmann:2000] overcomes this issue by
     I^\Lambda := \sum \limits_{\bsnu \in \Lambda} \zeta_{\Lambda, \bsnu} I^\bsnu, \qquad \zeta_{\Lambda, \bsnu} := \sum \limits_{\bse \in \{0,1\}^d : \bsnu+\bse \in \Lambda} (-1)^{|\bse|}.
 \end{equation*}
 
-Implementing this operator in a vectorized form suitable for high-performance computing is nontrivial, as vectorization requires inputs to conform to a uniform structure. However, each tensorized interpolant in \eqref{eq:ip_smolyak} involves reducing a higher-order tensor of unique shape $\bsnu$ via multiplication with one vector per dimension. A naive strategy would be to zero-pad all tensors in \eqref{eq:ip_smolyak} to the smallest common shape $(\max_{\bsnu \in \Lambda}(\nu_j))_{j=1}^d$. This, however, reintroduces the curse of dimensionality, as memory requirements grow exponentially with $d$.
[…]
