### 2025-08-14 #8505 @danielskatz
👋 @JoWestermann - I've proofread this, and suggest changes in https://github.com/JoWestermann/smolyax/pull/67. Please merge this, or let me know what you disagree with, then we can proceed.
