### 2025-08-14 ae85eb7
changes for JOSS publication (#67)
* Update paper.md

* Update paper.bib
