+++ paper/paper.bib
@@ -206,7 +206,7 @@ MRREVIEWER = {Brian\ Bradie},
 }
 
 @article{tate:2023,
-  title={UncertainSCI: A Python Package for Noninvasive Parametric Uncertainty Quantification of Simulation Pipelines},
+  title={UncertainSCI: A {P}ython Package for Noninvasive Parametric Uncertainty Quantification of Simulation Pipelines},
   author={Tate, Jess and Liu, Zexin and Bergquist, Jake A and Rampersad, Sumientra and White, Dan and Charlebois, Chantel and Rupp, Lindsay and Brooks, Dana H and MacLeod, Rob S and Narayan, Akil},
   journal={Journal of Open Source Software},
   volume={8},
