### 2025-08-14 a932bc9
Fixed two more typos in paper
