### 2025-08-14 #8505 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
