### 2025-08-14 #8505 @danielskatz
Congratulations to @JoWestermann (Josephine) and co-author on your JOSS publication!!

And thanks to @MoraruMaxim & @mjcarley for reviewing, and to @diehlpk for editing!!
JOSS depends on volunteers, and we couldn't do this without you
